pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "agent-job-finder"
version = "0.1.0"
description = "Autonomous AI agent that finds and applies to jobs using CrewAI, Ollama, and Supabase"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "crewai>=1.12.2",
    "cryptography>=46.0.7",
    "ddgs>=9.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "jupyter>=1.1.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "pdfplumber>=0.11.0",
    "pydantic-settings>=2.0.0",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.0.0",
    "reportlab>=4.4.10",
    "structlog>=24.0.0",
    "supabase>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "pre-commit>=4.0.0",
    "pip-audit>=2.7.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
# not strict — some 3rd-party libs lack stubs

[tool.pytest.ini_options]
testpaths = ["src/worker/tests"]
pythonpath = ["src"]
addopts = "-x -q"
markers = [
    "integration: full pipeline tests with mocked external I/O",
]
